+++ custom_typings/bower-json.d.ts
@@ -1,0 +1,3 @@
+declare module 'bower-json' {
+  export function read(path: string, options: any, callback: (err: any, json: any) => void): void;
+}