+++ custom_typings/bower-loger.d.ts
@@ -1,0 +1,4 @@
+declare module 'bower-logger' {
+  class BowerLogger extends NodeJS.EventEmitter {}
+  export = BowerLogger;
+}