+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - Upgrade `polymer-build` to `v0.5.0`, which means the build command will now use the new [`polymer-analyzer`](https://github.com/Polymer/polymer-analyzer)! See [the polymer-build changelog](https://github.com/Polymer/polymer-build/blob/v0.5.0/CHANGELOG.md) for more information.
 - `build`: Rename the `--include-dependencies` flag to `--extra-dependencies`
 - `polymer.json`: Rename the `includeDependencies` & `sourceGlobs` fields to `extraDependencies` & `sources`, respectively
+- Added an `install` command, which can install "variants" of Bower dependencies. TODO(justinfagnani): document variants.
 
 ## v0.17.0